feat(report): integrate ImageLightboxGallery for enhanced image display
Replaced the existing image rendering logic in ReportModal, ReviewStep, and ReportView components with the ImageLightboxGallery component. This update improves the user experience by providing a more interactive and visually appealing way to view images, while also simplifying the code by removing redundant error handling for broken images.
